+++ src/main/java/co/kr/petopia/service/ReplyService.java
@@ -8,13 +8,13 @@ import co.kr.petopia.vo.ReplyVO;
 public interface ReplyService {
 
     // 게시글 별 list
-    List<ReplyVO> getReplyList(Long content_idx);
+    List<ReplyVO> getReplyList(int content_idx);
 
     // 게시글 별 list + 페이징
-    List<ReplyVO> getReplyListPaging(Criteria cri, Long content_idx);
+    List<ReplyVO> getReplyListPaging(Criteria cri, int content_idx);
 
     // read
-    ReplyVO getReply(Long reply_idx);
+    ReplyVO getReply(int reply_idx);
 
     // insert
     int replyRegister(ReplyVO reply);
@@ -23,7 +23,7 @@ public interface ReplyService {
     int replyModify(ReplyVO reply);
 
     // delete
-    int replyRemove(Long reply_idx);
+    int replyRemove(int reply_idx);
 
 	ReplyVO getMyAnswerList(int content_idx);
